requirements-bronze-tools.txt
kafka-python==2.0.2
pyarrow==16.1.0
google-cloud-bigquery==3.25.0
